requirements.txt
anyio==3.3.3
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
fastapi==0.70.0
idna==3.3
pydantic==1.8.2
python-dotenv==0.19.1
requests==2.26.0
sniffio==1.2.0
soupsieve==2.2.1
starlette==0.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
